use a unique conversation for ssl in each direction
Previously a single Wireshark conversion was used for both directions
which lead to fragmented TLS messages being reassembled from both
directions. TLS fragments should only be reassembled across the same
direction.
Instead of a single conversation, we now use two, one for each
direction. Wireshark will automatically merge conversations with
swapped addresses and ports so to prevent that we combine both ports
into a single number, specify that as one of the ports, and use
`NO_PORT_B` to indicate that this is the only port.
This mirrors what we already do for reassembly of s7comm-plus messages,
except that we use a slightly different mapping to prevent overlap.